requirements.txt
openpyxl>=3.1.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
yfinance>=0.2.0
